Updating vimup script to include GUI

+++ Shell/pms/04-vimup.sh
@@ -16,7 +16,9 @@ function vimup {
   ./configure --prefix=/usr --enable-gui=gtk2
   make
   sudo checkinstall --maintainer="Brenton Horne \<brentonhorne77@gmail.com\>" --pkgversion="${_pkgver}"
-  rm $HOME/Programs/Deb/vim*.deb
+  if ! [[ -f $HOME/Programs/Deb/vim*.deb ]]; then
+    rm $HOME/Programs/Deb/vim*.deb
+  fi
   mv vim_${_pkgver}-1_amd64.deb $HOME/Programs/Deb
   cd $PWD
 }
